+++ spec/requests/admin_claim_checks_spec.rb
@@ -8,33 +8,31 @@ RSpec.describe "Admin claim approvals", type: :request do
       follow_redirect!
     end
 
-    describe "claim_approvals#create" do
+    describe "claim_checks#create" do
       let(:claim) { create(:claim, :submitted) }
 
       it "approves a claim" do
         freeze_time do
-          post admin_claim_approvals_path(claim_id: claim.id)
+          post admin_claim_checks_path(claim_id: claim.id, result: "approved")
 
           follow_redirect!
 
           expect(response.body).to include("Claim has been approved successfully")
 
-          claim.reload
-
-          expect(claim.approved_at).to eq(Time.zone.now)
-          expect(claim.approved_by).to eq("123")
+          expect(claim.check.checked_by).to eq("123")
+          expect(claim.check.result).to eq("approved")
         end
       end
 
-      context "when claim is unapprovable" do
+      context "when claim is already approved" do
         let(:claim) { create(:claim, :approved) }
 
         it "shows an error" do
-          post admin_claim_approvals_path(claim_id: claim.id)
+          post admin_claim_checks_path(claim_id: claim.id, result: "approved")
 
           follow_redirect!
 
-          expect(response.body).to include("Claim cannot be approved")
+          expect(response.body).to include("Claim already checked")
         end
       end
     end
@@ -47,11 +45,11 @@ RSpec.describe "Admin claim approvals", type: :request do
       follow_redirect!
     end
 
-    describe "claim_approvals#create" do
+    describe "claim_checks#create" do
       let(:claim) { create(:claim, :submitted) }
 
       it "does not allow a claim to be approved" do
-        post admin_claim_approvals_path(claim_id: claim.id)
+        post admin_claim_checks_path(claim_id: claim.id, result: "approved")
 
         expect(response.code).to eq("401")
       end